# docs/openapi.yaml
components:
  schemas:
...
    api.CreateReviewRequest:
      properties:
        pull_request_id:
          format: int64
          type: integer
      required:
        - pull_request_id
      type: object
...
    api.DiffReviewRequest:
      properties:
        diff_zip_base64:
          type: string
        repo_name:
          type: string
      required:
        - diff_zip_base64
        - repo_name
      type: object
...
    api.ErrorResponse:
      properties:
        error:
          type: string
      required:
        - error
      type: object
...
    api.ReviewResponse:
      properties:
        aiSummaryTitle:
          nullable: true
          type: string
        authorName:
          nullable: true
          type: string
        authorUsername:
          nullable: true
          type: string
        branch:
          nullable: true
          type: string
        commitHash:
          nullable: true
          type: string
        completedAt:
          format: date-time
          nullable: true
          type: string
        connectorId:
          format: int64
          nullable: true
          type: integer
        createdAt:
          format: date-time
          type: string
        friendlyName:
          nullable: true
          type: string
        id:
          format: int64
          type: integer
        metadata:
          additionalProperties: {}
          type: object
        mrTitle:
          nullable: true
          type: string
        orgId:
          format: int64
          type: integer
        prMrUrl:
          nullable: true
          type: string
        provider:
          nullable: true
          type: string
        repository:
          type: string
        startedAt:
          format: date-time
          nullable: true
          type: string
        status:
          type: string
        triggerType:
          type: string
        userEmail:
          nullable: true
          type: string
      required:
        - id
        - repository
        - status
        - triggerType
        - createdAt
        - orgId
      type: object
...
    api.ReviewsListResponse:
      properties:
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
        page:
          type: integer
        perPage:
          type: integer
        reviews:
          items:
            $ref: '#/components/schemas/api.ReviewResponse'
          type: array
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - perPage
        - totalPages
        - hasNext
        - hasPrevious
      type: object
...
    api.TriggerReviewRequest:
      properties:
        url:
          type: string
      required:
        - url
      type: object
...
info:
  title: LiveReview API
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /api/v1/connectors/trigger-review:
    post:
      description: TriggerReviewV2 handles the request to trigger a code review using the new decoupled architecture
      operationId: TriggerReviewV2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.TriggerReviewRequest'
      responses:
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      tags:
        - Connectors
...
  /api/v1/diff-review:
    post:
      description: |-
        DiffReview accepts a base64-encoded ZIP containing a unified diff and triggers a review.
        Authentication is handled by middleware. This handler creates the review record,
        marks it as processing, and enqueues the job for async execution by the worker.
      operationId: DiffReview
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.DiffReviewRequest'
      tags:
        - Diff-Review
...
  /api/v1/diff-review/{review_id}:
    get:
      description: GetDiffReviewStatus returns processing status or completed results for a diff review.
      operationId: GetDiffReviewStatus
      parameters:
        - in: path
          name: review_id
          required: true
          schema:
            type: string
      tags:
        - Diff-Review
...
  /api/v1/reviews:
    get:
      description: getReviews handles GET /api/v1/reviews with filtering and pagination
      operationId: getReviews
      parameters:
        - in: query
          name: page
          required: true
          schema:
            type: string
        - in: query
          name: per_page
          required: true
          schema:
            type: string
        - in: query
          name: status
          required: true
          schema:
            type: string
        - in: query
          name: provider
          required: true
          schema:
            type: string
        - in: query
          name: search
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ReviewsListResponse'
          description: OK
      tags:
        - Reviews
    post:
      description: |-
        createReview handles POST /api/v1/reviews. It resolves the PR/MR's web_url
        from the pull_requests table (org-scoped) and reuses TriggerReviewV2's
        phases unchanged (setupReviewContext -> prepareAuthentication ->
        createReviewRequest -> enrichMetadata -> trackActivity ->
        launchBackgroundProcessing), so both trigger entrypoints share identical
        authentication, LOC-quota, and background-processing behavior. The only
        addition is linking the created review back to its canonical PR/MR row via
        reviews.pull_request_id.
      operationId: createReview
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.CreateReviewRequest'
      responses:
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      tags:
        - Reviews
...
servers:
  - url: http://localhost:8888